Generate crx file using pure node.js, no need to exec shell scripts
Using the https://github.com/oncletom/grunt-crx npm module to package
the Chrome extension makes it possible to use on Windows systems.

There is a Grunt task (https://github.com/oncletom/grunt-crx) for this,
however it seems to be outdated and buggy, so this task is implemented
by manually calling to the `crx` module APIs.